requirements.txt
pyyaml>=6.0
requests>=2.31.0
langchain-core>=0.1.0
langchain-qdrant>=0.0.1
langchain-text-splitters>=0.0.1
qdrant-client>=1.5.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
pypdf>=3.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
streamlit>=1.25.0

